Add alert function

+++ pdf_front/mask/views.py
@@ -5,7 +5,7 @@ from django.views.generic import ListView, DetailView, View, FormView
 
 from .forms import PostForm
 from .models import Post
-from .utils import encrypt, check_pw
+from .utils import encrypt, check_pw, get_alert, set_alert
 
 
 def index(req, **kwargs):
@@ -21,43 +21,46 @@ class IndexView(ListView):
 
 
 class PostDetailView(View):
+
     def get(self, req, pk):
         context = {}
         try:
             post = Post.objects.get(pk=pk)
             context['post'] = post
-            context['msg'] = "INFO: 이미지를 보려면, 암호를 입력하셔야 합니다."
+            set_alert(req, "이미지를 보려면, 암호를 입력하셔야 합니다.", type="INFO")
+            return render(req, 'mask/check_password.html', context)
         except Exception:
-            context['msg'] = "WARN: The post was not found."
-        return render(req, 'mask/check_password.html', context)
+            set_alert(req, "The post was not found.", "ERROR", type="DANGER")
+            context['posts'] = Post.objects.order_by('-reg_date')
+
+            return render(req, 'mask/index.html', context)
 
     def post(self, req, pk):
         context = {}
 
         try:
             if 'password' not in req.POST.keys():
-                context['msg'] = "WARN: Error"
+                set_alert(req, "Error", "ERROR", "DANGER")
 
             input_pw = req.POST['password']
             post = Post.objects.get(pk=pk)
             context['post'] = post
-            print(req.POST)
             hashed_pw = post.password
-            print(input_pw)
             if check_pw(input_pw, hashed_pw):
-                context['msg'] = "INFO: Password match!"
-                print('check true')
+                # context['alert'] = get_alert("Password match!", type="SUCCESS")
+                if post.status == Post.Status.STAND_BY:
+                    set_alert(req, "작업 대기중입니다.", type="NORMAL")
+                else:
+                    set_alert(req, title="작업 완료!", msg="아래 다운로드 버튼을 눌러보세요!", type="SUCCESS")
 
                 return render(req, 'mask/detail.html', context)
             else:
-                print('check false')
-                context['msg'] = "WARN: Password doesn't match"
+                set_alert(req, "비밀번호가 틀립니다.", type="DANGER")
 
                 return render(req, 'mask/check_password.html', context)
         except Exception:
-            print(Exception)
-            context['msg'] = "WARN: The post was not found."
-        return render(req, 'mask/check_password.html', context)
+            set_alert(req, "The post was not found.", type="WARN")
+            return render(req, 'mask/index.html', context)
 
 
 class _PostDetailView(DetailView):
@@ -82,13 +85,39 @@ class PostFormView(FormView):
             return redirect('mask:detail', pk=pk)
 
 
-def delete_post(req, pk):
-    context = {}
-    try:
-        post = Post.objects.get(pk=pk)
-        post.delete()
-        context['msg'] = "INFO: Deleted."
-    except Exception:
-        context['msg'] = "WARN: The post was not found."
+class PostDeleteView(View):
+    def get(self, req, pk):
+        context = {}
+        try:
+            post = Post.objects.get(pk=pk)
+            context['post'] = post
+            set_alert(req, "이미지를 삭제하려면, 암호를 입력하셔야 합니다.", type="INFO")
+            return render(req, 'mask/check_password.html', context)
+        except Exception:
+            set_alert(req, "The post was not found.", "ERROR", type="DANGER")
+            context['posts'] = Post.objects.order_by('-reg_date')
 
-    return redirect('mask:index')
+            return render(req, 'mask/index.html', context)
+
+    def post(self, req, pk):
+        context = {}
+
+        try:
+            if 'password' not in req.POST.keys():
+                set_alert(req, "Error", type="DANGER")
+            input_pw = req.POST['password']
+            post = Post.objects.get(pk=pk)
+            context['post'] = post
+            hashed_pw = post.password
+            if check_pw(input_pw, hashed_pw):
+                post.delete()
+                set_alert(req, "삭제되었습니다.", type="SUCCESS")
+
+                return render(req, 'mask/index.html', context)
+            else:
+                set_alert(req, "비밀번호가 틀립니다.", type="DANGER")
+
+                return render(req, 'mask/check_password.html', context)
+        except Exception:
+            set_alert(req, "The post was not found.", type="WARN")
+            return render(req, 'mask/index.html', context)
